% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mspflights-package}
\alias{mspflights}
\alias{mspflights-package}
\title{mspflights: Flights Between MSP and LIT, STL from April 2013 to February 2015}
\description{
Airline on-time data for all flights departing from or arriving to MSP, LIT, and STL between April 2013 and February 2015. Also includes useful 'metadata' on airlines, airports, weather, and planes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jabenninghoff.github.io/mspflights/}
  \item \url{https://github.com/jabenninghoff/mspflights}
  \item Report bugs at \url{https://github.com/jabenninghoff/mspflights/issues}
}

}
\author{
\strong{Maintainer}: John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})

}
\keyword{internal}
